refactor bookkeeper
