+++ Gopkg.toml
@@ -50,3 +50,7 @@
 [[constraint]]
   name = "github.com/iotexproject/iotex-core"
   branch = "master"
+
+[[constraint]]
+  name = "github.com/spf13/cobra"
+  version = "0.0.3"
